requirements.txt
fastmcp
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
